requirements.txt
aiogram>=3.4.1
python-dotenv>=1.0.0
asyncpg>=0.29.0
aiohttp>=3.9.0
pytz>=2023.3